--- fe-0.6/pom.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <groupId>com.fernferret.allpay</groupId>
-        <artifactId>AllPay</artifactId>
-        <version>13</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>com.fernferret.allpay.economy</groupId>
-    <artifactId>fe-0.6</artifactId>
-    <version>1</version>
-    <packaging>pom</packaging>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.melonbrew</groupId>
-            <artifactId>Fe</artifactId>
-            <version>0.6</version>
-            <type>jar</type>
-        </dependency>
-    </dependencies>
-
-
-</project>
