--- itemecon/pom.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>AllPay</artifactId>
-        <groupId>com.fernferret.allpay</groupId>
-        <version>13</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>com.fernferret.allpay.economy</groupId>
-    <artifactId>itemecon</artifactId>
-    <version>1</version>
-    <packaging>pom</packaging>
-
-</project>
