+++ modules/commons/src/main/java/com/fernferret/allpay/commons/GenericBank.java
