+++ modules/commons/src/main/java/com/fernferret/allpay/commons/IncompleteBankException.java
