+++ modules/essentialsbank/src/main/java/com/fernferret/allpay/economy/EssentialsBank.java
