+++ modules/fe-0.7/pom.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <groupId>com.fernferret.allpay.economy</groupId>
+    <artifactId>fe-0.7</artifactId>
+    <version>1</version>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.melonbrew</groupId>
+            <artifactId>Fe</artifactId>
+            <version>0.7</version>
+            <type>jar</type>
+        </dependency>
+        <dependency>
+            <groupId>com.fernferret.allpay.commons</groupId>
+            <artifactId>commons</artifactId>
+            <version>1</version>
+        </dependency>
+    </dependencies>
+
+
+</project>
