+++ modules/multicurrency/src/main/java/com/fernferret/allpay/economy/MultiCurrencyBank.java
