+++ modules/pom.xml
@@ -1,0 +1,69 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <groupId>com.fernferret.allpay</groupId>
+    <artifactId>aggregate-dist</artifactId>
+    <version>1</version>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.fernferret.allpay.economy</groupId>
+            <artifactId>boseconomy</artifactId>
+            <version>1</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fernferret.allpay.economy</groupId>
+            <artifactId>econxp</artifactId>
+            <version>1</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fernferret.allpay.economy</groupId>
+            <artifactId>essentialsbank</artifactId>
+            <version>1</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fernferret.allpay.economy</groupId>
+            <artifactId>fe-0.6</artifactId>
+            <version>1</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fernferret.allpay.economy</groupId>
+            <artifactId>fe-0.7</artifactId>
+            <version>1</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fernferret.allpay.economy</groupId>
+            <artifactId>iconomy4</artifactId>
+            <version>1</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fernferret.allpay.economy</groupId>
+            <artifactId>iconomy5</artifactId>
+            <version>1</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fernferret.allpay.economy</groupId>
+            <artifactId>iconomy6</artifactId>
+            <version>1</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fernferret.allpay.economy</groupId>
+            <artifactId>itemecon</artifactId>
+            <version>1</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fernferret.allpay.economy</groupId>
+            <artifactId>multicurrency</artifactId>
+            <version>1</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fernferret.allpay.economy</groupId>
+            <artifactId>realeconomy</artifactId>
+            <version>1</version>
+        </dependency>
+    </dependencies>
+
+</project>
